requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-WTF>=1.2.0
discogs-client
python-dotenv>=1.0.0
pytz>=2023.3
redis>=4.5.0
psycopg2-binary>=2.9.0
requests-oauthlib>=1.3.1
schedule>=1.2.0
qrcode[pil]>=7.4.2
python-telegram-bot>=20.0